+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.167.2] - 2025-01-06 05:03:19
+
+### Fixed
+
+- Uncap New York real estate tax deduction.
+- Cap New York college tuition expenses credit and deduction per student.
+
 ## [1.167.1] - 2025-01-03 16:28:36
 
 ### Fixed
@@ -10500,6 +10507,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
+[1.167.2]: https://github.com/PolicyEngine/policyengine-us/compare/1.167.1...1.167.2
 [1.167.1]: https://github.com/PolicyEngine/policyengine-us/compare/1.167.0...1.167.1
 [1.167.0]: https://github.com/PolicyEngine/policyengine-us/compare/1.166.0...1.167.0
 [1.166.0]: https://github.com/PolicyEngine/policyengine-us/compare/1.165.0...1.166.0